Rename lambda names to be imperative
