+++ lambdas/build-ndjson/Dockerfile
