+++ lambdas/build-ndjson/requirements.txt
