+++ lambdas/db-write/hlss30_stac_example.ndjson
