--- lambdas/pgstac-loader/requirements.txt
@@ -1,4 +1,0 @@
-awslambdaric
-boto3
-aws-lambda-powertools
-pypgstac==0.4.5
